Tried putting make_test_tiles and assert_make_test_arrays_and_difference in conftest as fixtures but they didn't work as fixtures because fixtures don't handle arguments easily (as far as I can tell). So I tried just leaving those functions in conftest as non-fixture functions but test_deadwood_litter_rasterio.py couldn't find the two functions. So then I tried making a new file called test_helpers.py in test. When I ran the test with that, test_deadwood_litter_rasterio.py said there was no module named test_helpers. So I moved test_helpers.py to the project folder (outside test) and was able to load the functions fine. Testing works with this configuration but I don't like having the test helper functions outside test, and ideally they'd all be inside conftest.py (mixture of fixture and non-fixture functions).
